+++ core/include_internal/Prediction.h
@@ -10,14 +10,14 @@ namespace CmacLib
 	class Prediction : public IPrediction {
 	private:
 		std::vector<double> values;
-		std::vector<size_t> activeWeightIndices;
+		std::vector<unsigned int> activeWeightIndices;
 		std::vector<std::vector<double>> activeWeights;
 		std::vector<double> basisValues;
 		std::unique_ptr<IResult> result;
 
 	public: // setters
 		void SetValues(std::vector<double>& values);
-		void SetActiveWeightIndices(std::vector<size_t>& activeWeightIndices);
+		void SetActiveWeightIndices(std::vector<unsigned int>& activeWeightIndices);
 		void SetActiveWeights(std::vector<std::vector<double>>& weights);
 		void SetBasisValues(std::vector<double>& basis);
 		void SetResult(IResult* result);
@@ -28,7 +28,7 @@ namespace CmacLib
 
 	public: // override getters
 		const std::vector<double>& GetValues() override;
-		const std::vector<size_t>& GetActiveWeightIndices() override;
+		const std::vector<unsigned int>& GetActiveWeightIndices() override;
 		const std::vector<std::vector<double>>& GetActiveWeights() override;
 		const std::vector<double>& GetBasisValues() override;
 		IResult* const GetResult() override;
